% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{da_sefaz}
\alias{da_sefaz}
\title{da_sefaz}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26738 rows and 6 columns.
}
\usage{
da_sefaz
}
\description{
da_sefaz
}
\keyword{datasets}
